+++ src/App.tsx
@@ -1,0 +1,21 @@
+import { BrowserRouter, Routes, Route } from 'react-router-dom'
+import Layout from './components/Layout'
+import Home from './pages/Home'
+import Portfolio from './pages/Portfolio'
+import CV from './pages/CV'
+
+function App() {
+  return (
+    <BrowserRouter>
+      <Layout>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/portfolio" element={<Portfolio />} />
+          <Route path="/cv" element={<CV />} />
+        </Routes>
+      </Layout>
+    </BrowserRouter>
+  )
+}
+
+export default App